Вынес вспомогательные методы в отдельный файл
