+++ src/mock/task.js
@@ -1,11 +1,5 @@
+import {getRandomInteger} from '../utils.js';
 import {DEFAULT_REPEATING_DAYS,  DESCRIPTIONS, COLORS} from '../const.js';
-
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
 
 const generateDescription = () => {
   const randomIndex = getRandomInteger(0, DESCRIPTIONS.length - 1);
